requirements.txt
asgiref==3.3.0
certifi==2020.6.20
Django==3.1.3
django-leaflet==0.27.1
GDAL @ file:///C:/Users/GPAD/project/GDAL-3.1.4-cp38-cp38-win_amd64.whl
pathlib==1.0.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
pytz==2020.4
sqlparse==0.4.1
wincertstore==0.2
